Incorporated InfillDatesNeborsSets - Beta_03
- Removed save_step_vars_flag.
- out_add_info_df has only one column of n_nebors.
- Added n_recs column to out_add_info_df.
- Major changes to InfillStation class. It now takes a set of nebors
that are valid for the given infill dates only. So no need of finding
best nebors.
- NebSel is removed.
- Bug fix in DiscContVars. This happend when all stations said extreme
low or extreme high. The problem was in an if-else clause that happened
after computing the gy_arr. This is the problem that Thomas was having.
I thought that it was due to the sensitivity of the correlation matrix
but apperently it wasn't.